requirements.txt
ollama
sentence-transformers
google-generativeai
torch
torch-geometric
networkx
numpy
pydantic
python-dotenv
tqdm
scikit-learn
